requirements.txt
beautifulsoup4>=4.12
requests>=2.31
Pillow>=10.2
google-genai>=1.0
python-dotenv>=1.0
aiohttp>=3.9
numpy>=1.26
ruff>=0.9
